requirements.txt
python-dotenv
google-generativeai
streamlit
langchain
PyPDF2
chromadb
faiss-cpu
langchain-google-genai
langchain-community
pdf2image
youtube_transcript_api
pathlib